+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 <!-- SPDX-License-Identifier: Apache-2.0 -->
 # Changelog
 
+## [7.5.1.0] (Unreleased)
+
+### Features
+- [usd#2569](https://github.com/Autodesk/arnold-usd/issues/2569) - Support husk tokens for frame replacement
+- [usd#2435](https://github.com/Autodesk/arnold-usd/issues/2435) - Report USD warning and errors as Arnold warning and errors.
+
+### Bug Fixes
+
+- [usd#2572](https://github.com/Autodesk/arnold-usd/issues/2572) - Arnold product types not working if drivers:parameter:aov:name is not defined
+- [usd#2547](https://github.com/Autodesk/arnold-usd/issues/2547) - Changing material terminal is not updated with Hydra 2
+- [usd#2566](https://github.com/Autodesk/arnold-usd/issues/2566) - Fixed crashes when a render delegate is deleted while another is rendering
+- [usd#2568](https://github.com/Autodesk/arnold-usd/issues/2568) - Changing the root shader in IPR is not updated when it was present in the previous shading tree 
+
 ## [7.5.0.0] (Unreleased)
 
 ### Features
@@ -17,7 +30,23 @@
 - [usd#2530](https://github.com/Autodesk/arnold-usd/issues/2530) -  Support uniform and vertex-interpolated primvars that are indexed
 - [usd#2532](https://github.com/Autodesk/arnold-usd/issues/2532) -  Fix crashes in IPR with ArnoldGinstance primitives
 
-## [7.4.5.0] (Unreleased)
+## [7.4.5.2] (Unreleased)
+
+### Bug Fixes
+
+- [usd#2563](https://github.com/Autodesk/arnold-usd/issues/2563) - Incorrect region renders in hydra when one axis is not cropped
+- [usd#2558](https://github.com/Autodesk/arnold-usd/issues/2558) - Indexed user data shouldn't author the idxs companion to USD
+
+## [7.4.5.1] 2026-02-19
+
+### Bug Fixes
+
+- [usd#2557](https://github.com/Autodesk/arnold-usd/issues/2557) - Fix regression when a geometry has an instancer without any instances
+- [usd#2547](https://github.com/Autodesk/arnold-usd/issues/2547) - Changing material terminal is not updated with Hydra 2
+- [usd#2536](https://github.com/Autodesk/arnold-usd/issues/2536) - ArnoldUsd procedural filename is not reported as an asset
+- [usd#2561](https://github.com/Autodesk/arnold-usd/issues/2561) - Fix regression in husk with dataWindow NDC
+
+## [7.4.5.0] 2026-02-10
 
 ### Features
 
@@ -32,7 +61,7 @@
 - [usd#2471](https://github.com/Autodesk/arnold-usd/issues/2471) - Support for USD 25.11
 - [usd#2478](https://github.com/Autodesk/arnold-usd/issues/2478) - Use the Z AOV for the hydra depth buffer
 
-## [7.4.4.2] (Unreleased)
+## [7.4.4.2] 2026-01-22
 
 ### Bug Fixes